Fix building tests with Xcode 14 and benchmarks for iOS
Xcode 14 will no longer code-sign invalid application bundles, which includes
bundles with files that aren't where they're supposed to be. This means that we
need to actually put the resource files in the proper subdirectory of the
bundle. It also means that using the executable's directory as the working
directory when running the tests doesn't really work; we clean up correctly on
successly runs, but after a failed run there may be leftover files which makes
the next run of the tests break.

This happens to also cover nearly all of the changes required to support
running the object store tests and core benchmarks on iOS, so that's included
as well.